создает адаптеры, обновляет данные с сервера при eventEdit
