+++ src/controllers/point.js
@@ -2,6 +2,47 @@ import AbstractComponent from "../components/abstract-component.js";
 import EventComponent from "../components/event.js";
 import EventEditComponent from "../components/event-edit.js";
 import {RenderPosition, render, replace, remove} from "../utils/render.js";
+import PointModel from "../models/point.js";
+
+export const parseFormData = (formData, event, allOffers) => {
+  const selectedType = document.querySelector(`.event__type-output`)
+    .textContent.trim().split(` `)[0];
+
+  const checkedOffersTitle = Array.from(document.querySelectorAll(`.event__offer-checkbox`))
+      .filter((it) => it.checked)
+      .map((offer) => offer.getAttribute(`value`));
+
+  const getOffers = () => {
+    return allOffers.map((offer) => {
+      if (checkedOffersTitle.includes(offer.title)) {
+        return Object.assign({}, offer, {checked: true});
+      } else {
+        return Object.assign({}, offer, {checked: false});
+      }
+    });
+  };
+
+  const convertData = (dateString) => {
+    const monthIndex = dateString[3] + dateString[4];
+    const dayIndex = dateString[0] + dateString[1];
+    return monthIndex + `/` + dayIndex + dateString.slice(5);
+  };
+
+  return new PointModel({
+    "id": event.id,
+    "type": selectedType,
+    "offers": getOffers(),
+    "destination": Object.assign({}, event, {
+      name: formData.get(`event-destination`),
+      description: event.destination.description,
+      pictures: event.destination.pictures,
+    }),
+    "date_from": convertData(formData.get(`event-start-time`)),
+    "date_to": convertData(formData.get(`event-end-time`)),
+    "base_price": Number(formData.get(`event-price`)),
+    "is_favorite": !!(formData.get(`event-favorite`)),
+  });
+};
 
 export const Mode = {
   DEFAULT: `default`,
@@ -38,7 +79,7 @@ export default class PointController extends AbstractComponent {
     this._eventComponent = null;
     this._eventEditComponent = null;
     this._mode = Mode.DEFAULT;
-
+    this._offers = null;
   }
 
   _onEscKeyDown(evt) {
@@ -54,6 +95,7 @@ export default class PointController extends AbstractComponent {
   }
 
   render(event, mode) {
+    this._offers = event.eventType.offers;
     this._mode = mode;
     const oldEventComponent = this._eventComponent;
     const oldEventEditComponent = this._eventEditComponent;
@@ -79,8 +121,9 @@ export default class PointController extends AbstractComponent {
 
     this._eventEditComponent.setSubmitClickHandler((evt) => {
       evt.preventDefault();
-      const data = this._eventEditComponent.getData();
-
+      // debugger;
+      const formData = this._eventEditComponent.getData();
+      const data = parseFormData(formData, event, this._offers);
       this._onDataChange(this, event, data);
       // this._replaceEditToEvent();
     });
